File update

+++ Encode/BPSK.py
@@ -4,4 +4,4 @@ import torch
 def bpsk_modulator(codeword):
     codeword = 2 * codeword - 1
 
-    return codeword.to(torch.float)
+    return codeword
